research/arxiv_papers/MIA-Mind/requirements.txt
numpy
pandas
scikit-learn
tifffile
tqdm
mindspore-2.5.0